exposing the functionality to project a slant plane pixel into the ground plane latLong
